hicRouterに統一&Proxyの実装
